+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved memory usage when idle, including a case where memory could grow
   multiple GB when you have a large `/etc/hosts`.
 - Improved reliability of git refs syncing on concurrent operations.
+- The cli now errors immediately if the daemon is running a different version.
 
 ## [0.1.1] - 2026-08-10